refactor: move chunk merging code out of code_splitting.rs (#7285)
Closed #7282
