+++ crates/rolldown/src/stages/generate_stage/code_splitting.rs
@@ -22,7 +22,9 @@ use rolldown_utils::{
 };
 use rustc_hash::{FxHashMap, FxHashSet};
 
-use super::{GenerateStage, chunk_ext::ChunkCreationReason};
+use super::{
+  GenerateStage, chunk_ext::ChunkCreationReason, chunk_optimizer::try_insert_into_existing_chunk,
+};
 
 #[derive(Clone, Debug)]
 pub struct SplittingInfo {
@@ -740,7 +742,7 @@ impl GenerateStage<'_> {
     let mut pending_common_chunks: FxIndexMap<BitSet, Vec<ModuleIdx>> = FxIndexMap::default();
     // If it is allow to allow that entry chunks have the different exports as the underlying entry module.
     // This is used to generate less chunks when possible.
-    let allow_extension_optimize = (!matches!(self.options.preserve_entry_signatures, PreserveEntrySignatures::Strict)
+    let allow_optimize_chunk = (!matches!(self.options.preserve_entry_signatures, PreserveEntrySignatures::Strict)
         || !self.link_output.overrode_preserve_entry_signature_map.is_empty())
           // partial workaround of https://github.com/rolldown/rolldown/issues/5026#issuecomment-2990146735
           // TODO: maybe we could bailout peer chunk?
@@ -773,7 +775,7 @@ impl GenerateStage<'_> {
           chunk_id,
           self.link_output.metas[normal_module.idx].depended_runtime_helper,
         );
-      } else if allow_extension_optimize {
+      } else if allow_optimize_chunk {
         pending_common_chunks.entry(bits.clone()).or_default().push(normal_module.idx);
       } else {
         let mut chunk =
@@ -792,7 +794,7 @@ impl GenerateStage<'_> {
       }
     }
 
-    if allow_extension_optimize {
+    if allow_optimize_chunk {
       self.try_insert_common_module_to_exist_chunk(
         chunk_graph,
         bits_to_chunk,
@@ -829,7 +831,7 @@ impl GenerateStage<'_> {
         .filter(|idx| entry_chunk_idx.contains(idx))
         .collect_vec();
 
-      let item = Self::try_insert_into_exists_chunk(
+      let item = try_insert_into_existing_chunk(
         &chunk_idxs,
         &static_entry_chunk_reference,
         chunk_graph,
@@ -890,80 +892,6 @@ impl GenerateStage<'_> {
         }
       }
     }
-  }
-
-  fn try_insert_into_exists_chunk(
-    chunk_idxs: &[ChunkIdx],
-    entry_chunk_reference: &FxHashMap<ChunkIdx, FxHashSet<ChunkIdx>>,
-    chunk_graph: &ChunkGraph,
-    module_table: &ModuleTable,
-  ) -> Option<ChunkIdx> {
-    let mut user_defined_entry = vec![];
-    let mut dynamic_entry = vec![];
-    for &idx in chunk_idxs {
-      let Some(chunk) = chunk_graph.chunk_table.get(idx) else {
-        continue;
-      };
-      match chunk.kind {
-        ChunkKind::EntryPoint { meta, .. } => {
-          if meta.contains(ChunkMeta::UserDefinedEntry) {
-            user_defined_entry.push(idx);
-          } else {
-            dynamic_entry.push(idx);
-          }
-        }
-        ChunkKind::Common => return None,
-      }
-    }
-    let user_defined_entry_modules = Self::collect_entry_modules(&user_defined_entry, chunk_graph)?;
-
-    let merged_user_defined_chunk =
-      Self::find_merge_target(&user_defined_entry, &user_defined_entry_modules, module_table);
-    if !user_defined_entry.is_empty() {
-      let chunk_idx = merged_user_defined_chunk?;
-
-      let ret = dynamic_entry.iter().all(|idx| {
-        entry_chunk_reference
-          .get(&chunk_idx)
-          .map(|reached_dynamic_chunk| reached_dynamic_chunk.contains(idx))
-          .unwrap_or(false)
-      });
-      return ret.then_some(chunk_idx);
-    }
-
-    let dynamic_chunk_entry_modules = Self::collect_entry_modules(&dynamic_entry, chunk_graph)?;
-    Self::find_merge_target(&dynamic_entry, &dynamic_chunk_entry_modules, module_table)
-  }
-
-  /// Collects entry module indices from a list of chunk indices.
-  /// Returns `None` if any chunk is missing or has no entry module.
-  fn collect_entry_modules(
-    chunk_indices: &[ChunkIdx],
-    chunk_graph: &ChunkGraph,
-  ) -> Option<Vec<ModuleIdx>> {
-    let mut ret = Vec::with_capacity(chunk_indices.len());
-    for chunk_idx in chunk_indices {
-      let chunk = chunk_graph.chunk_table.get(*chunk_idx)?;
-      ret.push(chunk.entry_module_idx()?);
-    }
-    Some(ret)
-  }
-
-  /// Finds a chunk that can serve as the merge target for all entries.
-  /// A chunk can be the merge target if its entry module is imported by or equal to all other entry module.
-  fn find_merge_target(
-    chunk_indices: &[ChunkIdx],
-    entry_modules: &[ModuleIdx],
-    module_table: &ModuleTable,
-  ) -> Option<ChunkIdx> {
-    chunk_indices.iter().zip(entry_modules.iter()).find_map(|(chunk_idx, entry_module_idx)| {
-      let module = module_table[*entry_module_idx].as_normal().expect("Should be normal module");
-      let can_merge = entry_modules.iter().all(|other_entry_module_idx| {
-        *entry_module_idx == *other_entry_module_idx
-          || module.importers_idx.contains(other_entry_module_idx)
-      });
-      can_merge.then_some(*chunk_idx)
-    })
   }
 
   fn construct_static_entry_to_reached_dynamic_entries_map(
