+++ crates/rolldown/src/stages/generate_stage/mod.rs
@@ -59,6 +59,7 @@ use crate::{
 
 mod advanced_chunks;
 mod chunk_ext;
+mod chunk_optimizer;
 mod code_splitting;
 mod compute_cross_chunk_links;
 mod minify_chunks;
